requirement.txt
asn1crypto==0.24.0
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
firebase-admin==2.14.0
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
msgpack==0.6.0
pendulum==2.0.4
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.2
pycparser==2.19
python-dateutil==2.7.5
python-dotenv==0.10.1
python-telegram-bot==11.1.0
pytz==2018.7
pytzdata==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
